Move screen header to components
