+++ src/components/screen-header/LogoOutline.tsx
