+++ src/components/screen-header/full-header.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import {View} from 'react-native';
 import {useSafeAreaInsets} from 'react-native-safe-area-context';
-import ScreenHeader, {ScreenHeaderProps} from '.';
-import {StyleSheet} from '../theme';
+import ScreenHeader, {ScreenHeaderProps} from './index';
+import {StyleSheet} from '../../theme';
 
 export default function FullScreenHeader(props: ScreenHeaderProps) {
   const style = useHeaderStyle();
