+++ src/screens/Profile/Language/index.tsx
@@ -10,7 +10,7 @@ import {
   DEFAULT_LANGUAGE,
   appLanguages,
 } from '../../../translations';
-import ScreenHeader from '../../../ScreenHeader';
+import ScreenHeader from '../../../components/screen-header';
 const identity = (s: string) => s;
 export default function Language() {
   const {
